requirements.txt
anticaptchaofficial 1.0.59
beautifulsoup4      4.12.2
bs4                 0.0.1
Pillow              10.1.0
requests            2.31.0
selenium            4.10.0


Quando no windows, talvez seja necesário ChromeDriverManager da biblioteca webdriver_manager.chrome